react-with-flask/api/requirements.txt
flask
python-dotenv
requests
openai
numpy